Refactoring Update
scripts have been cut in more pieces, increasing the readability and making it more organized

Team decided to upgrade to Python 3.10 so we changed the code a little and put some basic match statements. We will implement more of them in the future.

Also I added a couple of sounds and music