Fixed the enemy (got stuck inside the walls)
Fixed the enemy choice of directions when encountering a wall. The move ability was wrongly assignated (for enemy only) due to the addition of diagonal points which were breaking the whole thing. Fixed it by adding the diagonals points only when the enemy is chasing the player (only time where he moves diagonally).